src/requirements.txt
numpy==1.24.3
pandas==2.0.2
requests==2.31.0
streamlit==1.24.1
python-dotenv==1.0.0
streamlit-extras==0.2.7
Pillow==10.0.0
pybase64==1.2.3